==> analysis/05_CommunityAnalysis.rmd <==
---
title: "Between Sample (Beta) Diversity of Microbes along a Salinity Gradient"
author: "Andrea Martinez"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.path = "../figures/05_CommunityAnalysis")
```

#Goals
1. Load in ohyloseq data with rooted tree
2. Evaluate sequencing depth and remove sample
3. Normalize the read counts between samples
4. Calculate community dissimilarities. Numbers between 0 and 1. If 0, completely similar versus if they are 1, then they're completely dissimilar.
  a. *Sorensen*: Shared Species as a binary value: Abundance-unweighted
  b. *Bray-Curtis*: Shared Abundant species: Abundance-weighted
  c. *(Abundance-) Weighted UNIFRAC*: Consider Abundant Species and where they fall on the tree
5. Visualize the community data with two unconstrained Ordinations:
 a. *PCoA*: Linear method. Eigenvalue = how much variation is explained by each axis. Choose to view axis 1,2,3, ect. and plot them together.
 b. *NMDS*: Non-linear. Smuch multiple Dimensions into 2 or 3. Need to report a stress value (ideally <0.15). 
6. Run statistics with PERMANOVA and betadispR.
 
 
# Setup

## Load Libraries
```{r load-libraries}
#install.packages("vegan")
pacman:: p_load(tidyverse, devtools, phyloseq, patchwork, vegan,
                install = FALSE)

# Load Station Colors
station_colors <- c(
  "Shipping Channel" = "dodgerblue4",
  "Aransas Bay" = "dodgerblue2",
  "Copano West" = "#D9CC3C",
  "Copano East" = "#A0E0BA",
  "Mesquite Bay" = "#00ADA7")
```

## Load Data
```{r load-physeq}
# Load in rooted phylogenetic tree
load("data/03_PhylogeneticTree/phytree_preprocessed_physeq.RData")
midroot_physeq_rm456
unrooted_physeq_rm456
```



# Explore Read Counts
# Raw Read Depth
```{r calc-seq-depth}
# Calculate the total number of reads per sample.
raw_TotalSeqs_df <- midroot_physeq_rm456 %>%
  # calculate the sample read sums
sample_sums() %>%
  data.frame()
# name the column
colnames(raw_TotalSeqs_df)[1] <- "TotalSeqs_df"
head(raw_TotalSeqs_df)

# make histogram of raw reads
raw_TotalSeqs_df %>%
  ggplot(aes(x = TotalSeqs_df)) +
  geom_histogram(bins = 50) +
  scale_x_continuous(limits = c(0,10000)) +
  labs(title = "Raw Sequencing Depth Distribution")+
  theme_classic()

```

##Remove lowly seq sample
```{r low-seq-sample}
raw_rooted_physeq <- midroot_physeq_rm456 %>%
  # remove lowly seq sample that was outlier in alpha diversity analysis
  subset_samples(names != "20210615-MA-ABB2F") %>%
  #any asvs unique to this sample will aso be removed
  prune_taxa(taxa_sums(.) >0, .)

# Inspect
raw_rooted_physeq

# what is the minimum number of sequences
 raw_rooted_physeq %>%
   sample_sums() %>%
   min()
```

# Normalize read counts
```{r scale-reads}
# Function to scale reads: http://deneflab.github.io/MicrobeMiseq/ 
# Scales reads by 
# 1) taking proportions
# 2) multiplying by a given library size of n
# 3) rounding 
# Default for n is the minimum sample size in your library
# Default for round is floor

matround <- function(x){trunc(x+0.5)}

scale_reads <- function(physeq, n = min(sample_sums(physeq)), round = "round") {
  
  # transform counts to n
  physeq.scale <- transform_sample_counts(physeq, function(x) {(n * x/sum(x))})
  
  # Pick the rounding functions
  if (round == "floor"){
    otu_table(physeq.scale) <- floor(otu_table(physeq.scale))
  } else if (round == "round"){
    otu_table(physeq.scale) <- round(otu_table(physeq.scale))
  } else if (round == "matround"){
    otu_table(physeq.scale) <- matround(otu_table(physeq.scale))
  }
  
  # Prune taxa and return new phyloseq object
  physeq.scale <- prune_taxa(taxa_sums(physeq.scale) > 0, physeq.scale)
  return(physeq.scale)
}
```

## Scale the reads and check the distribution of the seq depth
```{r scale-physeq}
min(sample_sums(raw_rooted_physeq))

# Scale reads by the above function
scaled_rooted_physeq <- 
  raw_rooted_physeq %>%
  scale_reads(round = "matround")

# Calculate the read depth
scaled_TotalSeqs_df <-
  scaled_rooted_physeq %>%
  sample_sums() %>%
  data.frame() %>%
colnames(scaled_TotalSeqs_df)[1] <- "TotalSeqs" 

#inspect
head(scaled_TotalSeqs_df)

# Check the range of the data
min_seqs <- min(scaled_TotalSeqs_df)
max_seqs <- max(scaled_TotalSeqs_df)

#range
max_seqs-min_seqs

# Plot Histogram
scaled_TotalSeqs_df %>%
  ggplot(aes(x = TotalSeqs)) +
  geom_histogram(bins = 50) +
  scale_x_continuous(limits = c(0,10000)) +
  labs(title = "Scaled Sequencing Depth at 2194")+
  theme_classic()
```


# Calculate & visualize community dissimilarity
Exploratory analyses from Paliy & Shankar (2016) paper, which is using unconstrained ordination methods like PCoA

## Soresen PCoA
```{r sorensen-pcoa}
# Calculate sorensen dissimilarity: Abundance-unweighted of shared taxa
scaled_soren_pcoa <- ordinate(
  physeq = scaled_rooted_physeq,
  method = "PCoA",
  distance = "bray", binary = TRUE)
#str(scaled_soren_pcoa)

# Plot the ordination 
plot_ordination(
  physeq = scaled_rooted_physeq,
  ordination = scaled_soren_pcoa,
  color = "station",
  title = "Sorensen PCoA") +
  scale_color_manual(values = station_colors) + 
  theme_bw()

# PERMANOVA of Sorensen
# make a new object that has sorensen dissimilarity matrix 
sorensen_distance <-
  phyloseq::distance(scaled_rooted_physeq, method = "bray", binary = TRUE)
str(sorensen_distance)

# metadata
metadata <-
  scaled_rooted_physeq %>%
  sample_data %>%
  data.frame()

# Actually run PERMANOVA
## Testing if the centroids of the data are similar or different?
adonis2(sorensen_distance ~ station, data = metadata)
```

## Bray-Curtis PCoA

## Weighted Unifrac PCoA

## Bray-Curtis NMDS

# Test for Statistical Significance with PERMANOVA & betadispR

# Session Information
For reproducibility

```{ r session information}
devtools::session_info
```
